+++ packages/react-components/src/hook/index.js
@@ -1,7 +1,9 @@
 import useOutsideClick from './use-outside-click'
+import useFontFaceObserver from './use-font-face-observer'
 
-export { useOutsideClick }
+export { useOutsideClick, useFontFaceObserver }
 
 export default {
   useOutsideClick,
+  useFontFaceObserver,
 }